% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counter.R
\docType{class}
\name{Counter}
\alias{Counter}
\title{Counter R6 Class}
\format{
An \href{https://r6.r-lib.org/index.html}{R6} object.
}
\description{
Counter R6 Class

Counter R6 Class
}
\details{
A simple counter object with increment and reset functionality.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{value}}{Current value of the counter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Counter-new}{\code{Counter$new()}}
\item \href{#method-Counter-increment}{\code{Counter$increment()}}
\item \href{#method-Counter-reset}{\code{Counter$reset()}}
\item \href{#method-Counter-get}{\code{Counter$get()}}
\item \href{#method-Counter-clone}{\code{Counter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Counter-new"></a>}}
\if{latex}{\out{\hypertarget{method-Counter-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Counter`` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$new(start = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start}}{Initial value of the counter (default = 0).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Counter-increment"></a>}}
\if{latex}{\out{\hypertarget{method-Counter-increment}{}}}
\subsection{Method \code{increment()}}{
Increment the counter by 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$increment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Counter-reset"></a>}}
\if{latex}{\out{\hypertarget{method-Counter-reset}{}}}
\subsection{Method \code{reset()}}{
Reset the counter to 0.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Counter-get"></a>}}
\if{latex}{\out{\hypertarget{method-Counter-get}{}}}
\subsection{Method \code{get()}}{
Get the current counter value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$get()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current value of the counter.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Counter-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Counter-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Counter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
