% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_operators.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Left-hand value.}

\item{y}{Right-hand value (returned if x is NULL).}
}
\value{
\code{x} if not NULL, otherwise \code{y}.
}
\description{
A lightweight null-coalescing operator that avoids a dependency on rlang.
Returns the left-hand side if not NULL, otherwise the right-hand side.
}
\examples{
NULL \%||\% "default"
"value" \%||\% "default"

}
