% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\docType{class}
\name{Logger}
\alias{Logger}
\title{Logger R6 Class}
\format{
An \href{https://r6.r-lib.org/index.html}{R6} object.
}
\description{
Logger R6 Class

Logger R6 Class
}
\details{
A simple logger that tracks messages and timestamps, optionally per Shiny session.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{session_id}}{Optional identifier for the session}

\item{\code{logs}}{A character vector of logged messages}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Logger-new}{\code{Logger$new()}}
\item \href{#method-Logger-log}{\code{Logger$log()}}
\item \href{#method-Logger-get_logs}{\code{Logger$get_logs()}}
\item \href{#method-Logger-clone}{\code{Logger$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-new"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-new}{}}}
\subsection{Method \code{new()}}{
Create a new `Logger`` object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$new(session_id = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{session_id}}{A string identifying the session or user (optional).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-log"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-log}{}}}
\subsection{Method \code{log()}}{
Add a message to the log with a timestamp.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$log(message)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{message}}{The message to log.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-get_logs"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-get_logs}{}}}
\subsection{Method \code{get_logs()}}{
Retrieve all logged messages.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$get_logs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A character vector of log entries.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Logger-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Logger-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Logger$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
