% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModDownload.R
\name{ModDownload}
\alias{ModDownload}
\title{ModDownload R6 Class}
\description{
Handles report generation and download with format selection (HTML/PDF).
Uses R6 for object-oriented Shiny module design.
}
\examples{
\dontrun{
# In app_server():
download <- ModDownload$new(id = "download")
download$server(inputs_r, perf_r)
}

}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ModDownload-initialize}{\code{ModDownload$new()}}
    \item \href{#method-ModDownload-server}{\code{ModDownload$server()}}
    \item \href{#method-ModDownload-clone}{\code{ModDownload$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModDownload-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ModDownload-initialize}{}}}
\subsection{\code{ModDownload$new()}}{
  Initialize a new ModDownload instance
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ModDownload$new(id = "download")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Character. Module namespace id.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModDownload-server"></a>}}
\if{latex}{\out{\hypertarget{method-ModDownload-server}{}}}
\subsection{\code{ModDownload$server()}}{
  Initialize server logic
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ModDownload$server(inputs_r, perf_r)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{inputs_r}}{Reactive list returned by ModInputs$server()}
      \item{\code{perf_r}}{Reactive tibble returned by ModOutputs$server()}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Called for side-effects; returns NULL invisibly
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModDownload-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModDownload-clone}{}}}
\subsection{\code{ModDownload$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ModDownload$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
