% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_outputs.R
\name{ModOutputs}
\alias{ModOutputs}
\title{ModOutputs R6 Class}
\description{
Main content area with KPI value boxes and 5 demo tabs showcasing different
tooltip/hover-info approaches. Uses R6 for object-oriented Shiny module design.
}
\examples{
\dontrun{
# In app_ui():
outputs <- ModOutputs$new(id = "outputs")
outputs$ui()

# In app_server():
outputs <- ModOutputs$new(id = "outputs")
perf_r <- outputs$server(inputs_r)
}

}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-ModOutputs-initialize}{\code{ModOutputs$new()}}
    \item \href{#method-ModOutputs-ui}{\code{ModOutputs$ui()}}
    \item \href{#method-ModOutputs-server}{\code{ModOutputs$server()}}
    \item \href{#method-ModOutputs-clone}{\code{ModOutputs$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModOutputs-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-ModOutputs-initialize}{}}}
\subsection{\code{ModOutputs$new()}}{
  Initialize a new ModOutputs instance
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ModOutputs$new(id = "outputs")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{id}}{Character. Module namespace id.}
    }
    \if{html}{\out{</div>}}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModOutputs-ui"></a>}}
\if{latex}{\out{\hypertarget{method-ModOutputs-ui}{}}}
\subsection{\code{ModOutputs$ui()}}{
  Build main content UI
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ModOutputs$ui()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{\link[shiny:tagList]{shiny::tagList()}} with KPI boxes and demo tabs
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModOutputs-server"></a>}}
\if{latex}{\out{\hypertarget{method-ModOutputs-server}{}}}
\subsection{\code{ModOutputs$server()}}{
  Initialize server logic
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ModOutputs$server(inputs_r)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{inputs_r}}{Reactive list returned by ModInputs$server()}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Reactive tibble with performance metrics
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModOutputs-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModOutputs-clone}{}}}
\subsection{\code{ModOutputs$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{ModOutputs$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
