% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_producer.R
\docType{class}
\name{OutputProducer}
\alias{OutputProducer}
\title{OutputProducer R6 Class}
\format{
An \href{https://r6.r-lib.org/index.html}{R6} object.
}
\description{
OutputProducer R6 Class

OutputProducer R6 Class
}
\details{
Generates tables, plots, and download handlers from an \code{InputCollector}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{collector}}{The \code{InputCollector} instance containing inputs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OutputProducer-new}{\code{OutputProducer$new()}}
\item \href{#method-OutputProducer-get_table}{\code{OutputProducer$get_table()}}
\item \href{#method-OutputProducer-get_plot}{\code{OutputProducer$get_plot()}}
\item \href{#method-OutputProducer-download_data}{\code{OutputProducer$download_data()}}
\item \href{#method-OutputProducer-clone}{\code{OutputProducer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputProducer-new"></a>}}
\if{latex}{\out{\hypertarget{method-OutputProducer-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{OutputProducer} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputProducer$new(collector)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{collector}}{An \code{InputCollector} object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputProducer-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-OutputProducer-get_table}{}}}
\subsection{Method \code{get_table()}}{
Produce a summary table of collected inputs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputProducer$get_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A \code{data.frame} with input names and values.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputProducer-get_plot"></a>}}
\if{latex}{\out{\hypertarget{method-OutputProducer-get_plot}{}}}
\subsection{Method \code{get_plot()}}{
Produce a ggplot based on numeric input values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputProducer$get_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A ggplot object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputProducer-download_data"></a>}}
\if{latex}{\out{\hypertarget{method-OutputProducer-download_data}{}}}
\subsection{Method \code{download_data()}}{
Create a download handler for the input table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputProducer$download_data(filename = "data.csv")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{Name of the downloaded file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{shiny::downloadHandler} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OutputProducer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OutputProducer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OutputProducer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
