% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactable_theme.R
\name{rsixer_reactable_theme}
\alias{rsixer_reactable_theme}
\title{Custom reactable theme for Rsixer}
\usage{
rsixer_reactable_theme()
}
\value{
A \code{reactable::reactableTheme()} object
}
\description{
Creates a dark theme for reactable tables matching the Rsixer terminal aesthetic.
}
