% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\title{App UI}
\usage{
app_ui()
}
\description{
Application UI for Rsixer.
}
\details{
The UI provides buttons for the \code{Counter} demo as well as inputs for
collecting text, numeric, and file values. A table, plot, and download
button display the outputs generated from these inputs.
}
